Add erasing arrow function and save to local storage
Delete backspace arrow works now, it storages the values temporarily, then it determines when to change from erasing a value to erasing a digit.
The information of past operations is now stored in localstorage. it comes with date stamp to use it as reference later.
Remove most unnecessary console logs.