Setting Apostrophe validator
If deleteArrow() detects movement is a symbol, it will trigger an apostrophe that will indicate that the currentSign can be changed, it won't affect the value of the current operation if the user still wants to execute.
deleteArrow() used to rid currentSign at all, not being able for user to continue with the operation and deleting the whole value of currentSign, this was corrected already.